requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.3
beautifulsoup4>=4.12.3 